# Speed up Iterator parsers by swapping out Vector for Circular Buffer