@vovapolu 2016-08-28
Do we want to update a readme benchmarks after this PR?