@lihaoyi 2016-08-28
Sure, I'll update the benchmarks, but let's do so as a separate PR. I'm continuing to poke around and tweak things and that will probably have other perf impact in future